Decompose sema module into focused submodules
Reduce mod.rs from 924 to 155 lines by extracting types and methods
into dedicated submodules with single responsibilities:

- info.rs: FunctionInfo, MethodInfo, ConstInfo, AnonMethodSig types
- gather.rs: GatherOutput struct for declaration gathering output
- output.rs: AnalyzedFunction, SemaOutput result types
- inference_ctx.rs: InferenceContext for type inference
- visibility.rs: is_accessible, resolve_enum_through_module
- imports.rs: evaluate_const_initializers, import resolution
- anon_structs.rs: find_or_create_anon_struct
- sema_ctx_builder.rs: build_sema_context, inference building
- file_paths.rs: FileId to path mapping helpers

This is a pure refactoring with no functionality changes. All existing
tests pass and the public API is preserved via re-exports.

Fixes rue-q7vi
