+++ crates/rue-air/src/sema/inference_ctx.rs
@@ -1,0 +1,36 @@
+//! Pre-computed type information for constraint generation.
+//!
+//! This module contains [`InferenceContext`] which holds function, struct, enum,
+//! and method signature maps in `InferType` format for use in Hindley-Milner
+//! type inference.
+
+use std::collections::HashMap;
+
+use lasso::Spur;
+
+use crate::inference::{FunctionSig, MethodSig};
+use crate::types::{StructId, Type};
+
+/// Pre-computed type information for constraint generation.
+///
+/// This struct holds the function, struct, enum, and method signature maps
+/// converted to `InferType` format for use in Hindley-Milner type inference.
+/// Building this once and reusing it for all function analyses avoids the
+/// O(n²) cost of rebuilding these maps for each function.
+///
+/// # Performance
+///
+/// For a program with 100 functions and 50 structs:
+/// - **Before**: 100 × (HashMap rebuild + InferType conversions) per analysis
+/// - **After**: 1 × (HashMap build + InferType conversions) total
+#[derive(Debug)]
+pub struct InferenceContext {
+    /// Function signatures with InferType (for constraint generation).
+    pub func_sigs: HashMap<Spur, FunctionSig>,
+    /// Struct types: name -> Type::Struct(id).
+    pub struct_types: HashMap<Spur, Type>,
+    /// Enum types: name -> Type::Enum(id).
+    pub enum_types: HashMap<Spur, Type>,
+    /// Method signatures with InferType: (struct_id, method_name) -> MethodSig.
+    pub method_sigs: HashMap<(StructId, Spur), MethodSig>,
+}
